Time: 5 ms (17.09%), Space: 6.3 MB (26.79%) - LeetHub

+++ 1332-remove-palindromic-subsequences/1332-remove-palindromic-subsequences.cpp
@@ -3,7 +3,7 @@ public:
     int removePalindromeSub(string s) {
         if (s.length() == 0) return 0;
 
-        int i = 0, j = s.length() - 1;
+        short i = 0, j = s.length() - 1;
         while (i < j) {
             if (s[i] == s[j]) {
                 i++;
